Types

+++ types.go
@@ -67,11 +67,11 @@ type (
 	// ScrapeAhrefs represents the metrics retrieved from the
 	// Ahrefs API including cost, rows and if it was cached.
 	ScrapeAhrefs struct {
-		DR       float64 `json:"dr" bson:"dr"` // Domain Ranking
-		Rank     *int    `json:"rank" bson:"rank"`
-		RowsUsed int     `json:"rows_used" bson:"rows_used"`
-		Cost     float64 `json:"cost" bson:"cost"`
-		Cached   bool    `json:"cached" bson:"cached"`
+		DR       float64  `json:"dr" bson:"dr"` // Domain Ranking
+		Rank     *float64 `json:"rank" bson:"rank"`
+		RowsUsed int      `json:"rows_used" bson:"rows_used"`
+		Cost     float64  `json:"cost" bson:"cost"`
+		Cached   bool     `json:"cached" bson:"cached"`
 	}
 	// BackLinkCheck represents the data sent to the Lambda function
 	// for checking if a backlink appears on the page.
